+++ src/Components/ListRepo/index.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import List from '@material-ui/core/List';
+
+
+export default function ListRepo({itemList}) {
+
+
+    return (
+        <>
+            <List dense={true}>
+                {itemList}
+            </List>
+        </>
+    )
+}
